Небольшой рефакторинг

+++ DataAccessLayer/Models/User.cs
@@ -5,9 +5,9 @@ namespace DataAccessLayer.Models;
 
 public partial class User
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 
-    public int RoleId { get; set; }
+    public long RoleId { get; set; }
 
     public string Email { get; set; } = null!;
